chore(): update storybook examples
