+++ packages/react-ui/components/DateRangePicker/__docs__/DateRangePicker.docs.stories.tsx
@@ -11,10 +11,15 @@ import {
   LocaleContext,
   Select,
 } from '@skbkontur/react-ui';
-import { ThemeIn } from 'react-ui/lib/theming/Theme';
 
 import { Meta, Story } from '../../../typings/stories';
-import { Validations as ValidationsStory } from '../__stories__/DateRangePicker.stories';
+import {
+  Component,
+  CustomChildrenVertical,
+  CustomChildrenWithoutDash,
+  Sizes,
+  Validations as ValidationsStory,
+} from '../__stories__/DateRangePicker.stories';
 
 export default {
   title: 'Date Components/DateRangePicker',
@@ -22,36 +27,18 @@ export default {
   parameters: { creevey: { skip: true } },
 } as Meta;
 
-export const Example1: Story = () => {
-  const [value, setValue] = React.useState(['', '']);
-  const minDate = '08.07.2024';
-  const maxDate = '18.08.2024';
-
-  return (
-    <DateRangePicker
-      value={value}
-      minDate={minDate}
-      maxDate={maxDate}
-      onValueChange={([start, end]) => setValue([start, end])}
-    >
-      <DateRangePicker.Start />
-      <DateRangePicker.Separator />
-      <DateRangePicker.End />
-    </DateRangePicker>
-  );
-};
+export const Example1 = Component;
 Example1.storyName = 'Выбор периода';
 
 /**
  * Для валидаций используйте `DateRangePicker.validate(value, options)`, который принимает:
  * - `value` — проверяемые значения `['dd.mm.yyyy', 'dd.mm.yyyy']`
  * - `options` — объект с настройками `{ startOptional, endOptional, minDate, maxDate }`
- * 
- * Возвращается объект валидацией полей start и end в формате `[true, false]`
+ *
+ * Возвращается валидация полей start и end в формате `[true, false]`
  */
 export const Validations = ValidationsStory;
 Validations.storyName = 'Валидации';
-
 
 /**
  * Пример с кастомизацией темы и кастомным рендером дня
@@ -124,43 +111,10 @@ export const ExampleCustomOptional: Story = () => {
 };
 ExampleCustomOptional.storyName = 'Открытые даты начала или конца';
 
-
 /**
  * Через параметр `size` доступны размеры `large`, `medium` и `small`. С помощью токенов `calendar*` можно управлять размерами календаря
  */
-export const ExampleSizes: Story = () => {
-  const [valueS, setValueS] = React.useState(['', '']);
-  const [valueM, setValueM] = React.useState(['', '']);
-  const [valueL, setValueL] = React.useState(['', '']);
-
-  const theme = React.useContext(ThemeContext);
-  const createTheme = (tokens: ThemeIn) => ThemeFactory.create(tokens, theme);
-
-  return (
-    <Gapped vertical gap={16}>
-      <ThemeContext.Provider value={createTheme({ calendarCellWidth: '44px', calendarCellHeight: '44px' })}>
-        <DateRangePicker size="large" value={valueL} onValueChange={setValueL}>
-          <DateRangePicker.Start />
-          <DateRangePicker.Separator />
-          <DateRangePicker.End />
-        </DateRangePicker>{' '}
-      </ThemeContext.Provider>
-
-      <ThemeContext.Provider value={createTheme({ calendarCellWidth: '36px', calendarCellHeight: '36px' })}>
-        <DateRangePicker size="medium" value={valueM} onValueChange={setValueM}>
-          <DateRangePicker.Start />
-          <DateRangePicker.Separator />
-          <DateRangePicker.End />
-        </DateRangePicker>
-      </ThemeContext.Provider>
-      <DateRangePicker size="small" value={valueS} onValueChange={setValueS}>
-        <DateRangePicker.Start />
-        <DateRangePicker.Separator />
-        <DateRangePicker.End />
-      </DateRangePicker>
-    </Gapped>
-  );
-};
+export const ExampleSizes = Sizes;
 ExampleSizes.storyName = 'Размеры';
 
 /**
@@ -169,34 +123,11 @@ ExampleSizes.storyName = 'Размеры';
  * - `<DateRangePicker.End />`
  * - `<DateRangePicker.Separator />`
  */
-export const ExampleCustomWithoutDash: Story = () => {
-  const [value, setValue] = React.useState(['', '']);
-  return (
-    <DateRangePicker value={value} onValueChange={setValue}>
-      <DateRangePicker.Start style={{ borderRadius: 0 }} />
-      <DateRangePicker.End style={{ marginLeft: -1, borderRadius: 0 }} />
-    </DateRangePicker>
-  );
-};
+export const ExampleCustomWithoutDash = CustomChildrenWithoutDash;
 ExampleCustomWithoutDash.storyName = 'Поля без тире';
 
-export const ExampleCustomVertical: Story = () => {
-  const [value, setValue] = React.useState(['', '']);
-
-  return (
-    <DateRangePicker value={value} onValueChange={([start, end]) => setValue([start, end])}>
-      <Gapped gap={4} vertical>
-        <DateRangePicker.Start />
-        <DateRangePicker.End />
-      </Gapped>
-    </DateRangePicker>
-  );
-};
+export const ExampleCustomVertical = CustomChildrenVertical;
 ExampleCustomVertical.storyName = 'Вертикальное расположение';
-
-/**
- *
- */
 
 export const ExampleDateFormat: Story = () => {
   const [value, setValue] = React.useState(['', '']);
